Clean up classes
